requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
openai==1.30.1
sqlalchemy==2.0.30
pydantic==2.7.1
passlib[bcrypt]==1.7.4 # Добавлено для хеширования паролей
python-multipart==0.0.9 # Добавлено для поддержки форм FastAP
httpx==0.27.0  # НОВОЕ: Явно добавляем httpx и указываем версию.
python-jose[cryptography]
bcrypt